[qml] Fix segfault when resizing video widget with stopped player
When resizing QML engine invokes paint() that draws frame.
If player didn't started after launch it has not frame yet and app
crashes.
If player stopped it draws last frame as in paused state.

To fix it paint() must consider player state.
