+++ src/qml/QmlVideoPlayer.cpp
@@ -30,30 +30,14 @@
 VlcQmlVideoPlayer::VlcQmlVideoPlayer(QQuickItem *parent)
     : VlcQmlVideoObject(parent),
       _hasMedia(false),
-      _instance(NULL),
-      _player(NULL),
-      _media(NULL),
-      _audioManager(NULL),
       _autoplay(true),
       _deinterlacing(Vlc::Disabled),
       _seekable(true)
 {
-    _instance = new VlcInstance(VlcCommon::args(), this);
-    _instance->setUserAgent(qApp->applicationName(), qApp->applicationVersion());
-    _player = new VlcMediaPlayer(_instance);
-    connect(_player, SIGNAL(stateChanged()), this, SIGNAL(stateChanged()));
-    connect(_player, SIGNAL(seekableChanged(bool)), this, SLOT(s_seekableChanged(bool)));
-    _audioManager = new VlcAudio(_player);
 }
 
 VlcQmlVideoPlayer::~VlcQmlVideoPlayer()
 {
-    _player->stop();
-
-    delete _audioManager;
-    delete _media;
-    delete _player;
-    delete _instance;
 }
 
 void VlcQmlVideoPlayer::close()
